+++ test/unit/tools/cane.rb
@@ -4,12 +4,10 @@ module Test
       # Test for the 'cane' tool within the quality gem
       module Cane
         def cane_expected_args
-          "-f '{Rakefile,{*,.*}.{rb,rake,gemspec}," \
-          '{src,app,config,db,lib,test,spec,feature}/**/' \
-          "{*,.*}.{rb,rake,gemspec}}' " \
-          "--abc-exclude '{db/schema.rb}' " \
-          "--style-exclude '{db/schema.rb}' " \
-          "--doc-exclude '{db/schema.rb}'"
+          "-f '#{expected_ruby_source_glob}' " \
+          "--abc-exclude '{**/vendor/**,db/schema.rb}' " \
+          "--style-exclude '{**/vendor/**,db/schema.rb}' " \
+          "--doc-exclude '{**/vendor/**,db/schema.rb}'"
         end
 
         def expect_cane_run(quality_checker)